run `python -u` to unbuffer output
By default python output is line buffered when stdout is a terminal and
fully buffered when it's not.  This can be surprising when you redirect
the output to a file, and new data doesn't appear in a timely fashion.
Running python with `-u` make output unbuffered, which is less surprising.